Backend/requirements.txt
deepface==0.0.93
numpy==2.0.2
pillow==11.0.0
tensorflow==2.18.0
keras==3.7.0
opencv-python==4.10.0.84
mtcnn==1.0.0
requests==2.32.3
tf_keras==2.18.0
retina-face==0.0.17
scipy==1.12.0
cloudinary==1.37.3